% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated_model.R
\name{create_mult}
\alias{create_mult}
\title{Create a mult object which can be fit using the zipvb function.}
\usage{
create_mult(vy, mX, mZ, sigma2.beta, m = ncol(mZ), blocksize = 1,
  spline_dim = 0, v = 0)
}
\arguments{
\item{vy}{The vector of responses.}

\item{mX}{The matrix of fixed effects covariates.}

\item{mZ}{The matrix of random effects covariates.}

\item{sigma2.beta}{The variance of the fixed effects, beta.}

\item{m}{The number of groups of random effects.}

\item{blocksize}{The block size of the random effects e.g. 1 for random intercepts, 2 for random slopes.}

\item{spline_dim}{The dimension of the splines}

\item{v}{The degrees of freedom of the Inverse WIshart distribution}
}
\value{
A mult object, initialised to contain all of the parameters.
}
\description{
Create a mult object which can be fit using the zipvb function.
}

