% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipvb.R
\docType{package}
\name{zipvb}
\alias{zipvb}
\alias{zipvb-package}
\title{Package-level documentation}
\description{
Fit Bayesian zero-inflated Poisson regression models using the Gaussian Variational
					 appropximation.
}

