% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated_model (Mark Greenaway's
%   conflicted copy 2017-08-13).R, R/zero_inflated_model.R, R/zipvb.R
\docType{package}
\name{zipvb}
\alias{zipvb}
\alias{zipvb-package}
\title{Fit a zero-inflated Poisson model using Variational Bayes/Gaussian Variational Approximation.}
\usage{
zipvb(mult, method = "gva", verbose = FALSE,
  plot_lower_bound = FALSE, glm_init = TRUE)

zipvb(mult, method = "gva", verbose = FALSE,
  plot_lower_bound = FALSE, glm_init = TRUE)
}
\arguments{
\item{mult}{A mult object, created by create_mult.}

\item{method}{The algorithm used for fitting the model. One of "laplace", "gva", "gva2",
"gva_nr".}

\item{verbose}{A logical flag, for controlling the amount of debugging output.}

\item{plot_lower_bound}{A logical flag, for controlling whether the lower bound is plotted
as the fitting algorithm is being run.}

\item{glm_init}{A logical flag, for controlling whether the model fitting algoritym is
initialised using co-efficients from a Poisson model fit using glm.}

\item{mult}{A mult object, created by create_mult.}

\item{method}{The algorithm used for fitting the model. One of "laplace", "gva", "gva2",
"gva_nr".}

\item{verbose}{A logical flag, for controlling the amount of debugging output.}

\item{plot_lower_bound}{A logical flag, for controlling whether the lower bound is plotted
as the fitting algorithm is being run.}

\item{glm_init}{A logical flag, for controlling whether the model fitting algoritym is
initialised using co-efficients from a Poisson model fit using glm.}
}
\value{
A list detailing the model fit.

A list detailing the model fit.
}
\description{
Fit Bayesian zero-inflated Poisson regression models using the Gaussian Variational
					 appropximation.
}
